auto-generate html electrons
