+++ src/style/atom-animation.css
@@ -167,7 +167,7 @@
     display: flex;
     align-items: center;
 }
-.noble-gas-nucleus      { background-color: blueviolet !important; }
+.noble_gas-nucleus      { background-color: blueviolet !important; }
 .alkali-metal-nucleus   { background-color: orange !important; }
 .transition-metal-nucleus { background-color: rgb(255, 230, 0) !important; }
 .unknown-nucleus { background-color: rgb(32, 32, 32) !important; }
